crcX: Fix yield condition
Yield check was always compared to the same value, move the _count
increment inside the loop to correct this behaviour. And since this
makes the add(arr, len) code nearly identical to add(val) move the check
to add(val) and call it in add(arr, val) instead of duplicating code.